Preserving Creation Date
Need to preserve created_at date for the entry so if it were to be undeleted it will not show up as a recently added entry. When undeleting an entry, created_at timestamp will need to be transferred back to videos table. updated_at timestamp will be used to track changes, as intended.