HTTP Content Populate
Updated scripts to grab title, description and thumbnail image from the source site whenever needed. If the title field of an entry is left blank it triggers an operation to fetch information via GET request from the source site. Using regular expressions stored in the Sites table different pieces of information are extracted from the source and populated in the Videos table.